feat: default handler now accepts _every_ msg type
allow downstream integrations to use the default handler to filter out
all the different message types they might want to use. this way we
needn't create `setLocationHandler`, `setCallHandler`, etc.

this also modifies the exports of this library to add the `MESSAGE_TYPE`
map to the WickrBot class. it's not the most elegant way of handling
this, but with the way we're currently doing the exports it is the only
way to export > 1 thing afaik. changing the import behavior is something
for a v2.